+++ smacc_client_library/move_base_z_client/move_base_z_client_plugin/src/components/waypoints_navigator/waypoints_navigator.cpp
@@ -141,7 +141,7 @@ void WaypointNavigator::loadWayPointsFromFile(std::string filepath)
 
     if (wp_node != NULL)
     {
-      for (unsigned int i = 0; i < wp_node->size(); ++i)
+      for (uint64_t i = 0; i < wp_node->size(); ++i)
       {
         // Parse waypoint entries on YAML
         geometry_msgs::Pose wp;
